+++ src/kigl/GLBufferRange.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ki/GL.h"
+#include "kigl/kigl.h"
 
 struct GLBufferRange {
     size_t m_maxCount = 0;
@@ -33,7 +33,7 @@ struct GLBufferRange {
         return m_entrySize * m_usedCount;
     }
 
-    inline size_t getLengthFor(int count) const {
+    inline size_t getLengthFor(size_t count) const {
         return m_entrySize * count;
     }
